requirements.txt
# Azure SDK dependencies for management lock operations
azure-mgmt-resource>=21.0.0
azure-identity>=1.12.0
azure-core>=1.26.0
